#40: thredds crawler is crawling - remains to see if it works correctly. This is time consuming...
